Move Validation
Working on being able to get all possible moves for all possible pieces on the board. Also removed the square class. Piece class absorbed all of its additional data members. Havent fully finished move validation for pawn as we will need to add en passant as well as make it so that it can only move to diagonally if there is a piece of opposite color present in the diagonal squares. Next objective to complete (after pawn move validation) is making a function to actually move the pieces. this will involve a functionality to capture. with moving a piece, we will also need to make sure that moving that piece wont expose the king to an attack.