Further Updates
Player/computer can now move pieces. Simple validation currently happens for the move (i.e. check moving correct color, spot to movie piece in has piece, etc.). Also added an error.py to hold all custom error classes. Next is to check if user is trying to castle, and implement en passant functionality, as well as invalidate a move and reverse it if it causes a check as well as raise an error for the move.