Castling + Checking if Move Kills King
Made a function that will allow the user to make a move then it will switch to the other colors turn. Also implemented the functionalities for castling. Finally, also added the ability to check if a move will kill the king. This means either you move a piece and expose your king to an attack or your king is currently in check and your attempted move will not take it out of check. Final functionalities required to implement are en passant, checkmate, and stalemate. stalemate will include no valid moves for a color as well as an insufficient material stalemate.